BUGFIX: Reading Cloudinary uploader API key from environment variables (#13)
Discovered during testing that the Create Recipe modal was not allowing us to upload images which prevented the recipe from being created. Testing on the codebase prior to refactoring the reading of the environment variables works, so we are going to re-introduce the `dotenv` module's reading of the vars back into the two files from which they were removed.